move luawindow to its own module
